+++ R/explore.R
@@ -1,46 +1,3 @@
-#' LIME: sampling for local exploration by changing one value per observation.
-#'
-#' @param data Data frame from which observations will be generated.
-#' @param explained_instance A row in an original data frame (as a data.frame).
-#' @param size Number of observations to be generated.
-#' @param fixed_variables Names of column which will not be changed while sampling.
-#'
-#' @return data.frame
-#'
-
-generate_neighbourhood2 <- function(data, explained_instance, size, fixed_variables) {
-  data <- data.table::as.data.table(data)
-  neighbourhood <- data.table::rbindlist(lapply(1:size, function(x) explained_instance))
-  for(k in 1:nrow(neighbourhood)) {
-    picked_var <- sample(1:ncol(data), 1)
-    data.table::set(neighbourhood, i = as.integer(k), j = as.integer(picked_var),
-                    data[sample(1:nrow(data), 1), picked_var, with = FALSE])
-  }
-  as.data.frame(set_constant_variables(neighbourhood, explained_instance, fixed_variables))
-}
-
-
-#' LIME: sampling for local exploration by permuting all columns.
-#' 
-#' @param data Data frame from which observations will be generated.
-#' @param explained_instance A row in an original data frame (as a data.frame).
-#' @param size Number of observations to be generated.
-#' @param fixed_variables Names of column which will not be changed while sampling.
-#'
-#' @return data frame
-#'   
-
-permutation_neighbourhood <- function(data, explained_instance, size, fixed_variables) {
-  neighbourhood <- data.table::rbindlist(lapply(1:size, function(x) 
-                                                          explained_instance))
-  for(k in 1:ncol(neighbourhood)) {
-    data.table::set(neighbourhood, j = as.integer(k), 
-                    value = data[sample(1:nrow(data), size, replace = TRUE), 
-                                 k])
-  }
-  set_constant_variables(neighbourhood, explained_instance, fixed_variables)
-}
-
 #' Generate dataset for local exploration.
 #'
 #' @param data Data frame from which new dataset will be simulated.
@@ -50,9 +7,11 @@ permutation_neighbourhood <- function(data, explained_instance, size, fixed_vari
 #' @param size Number of observations is a simulated dataset.
 #' @param method If "live", new observations will be created by changing one value
 #'        per observation. If "lime", new observation will be created by permuting  all
-#'        columns of data.
+#'        columns of data. If "normal", numerical features will be sampled from multivariate
+#'        normal distribution specified by ... arguments mu and Sigma.
 #' @param fixed_variables names or numeric indexes of columns which will not be changed
 #'        while sampling.
+#' @param ... Mean and covariance matrix for normal sampling method.
 #'
 #' @return list consisting of
 #' \item{data}{Simulated dataset.}
@@ -63,7 +22,7 @@ permutation_neighbourhood <- function(data, explained_instance, size, fixed_vari
 #'
 #' @examples
 #' \dontrun{
-#' dataset_for_local_exploration <- sample_locally(data = wine,
+#' dataset_for_local_exploration <- sample_locally2(data = wine,
 #'                                                explained_instance = wine[5, ],
 #'                                                explained_var = "quality",
 #'                                                size = 50,
@@ -72,21 +31,27 @@ permutation_neighbourhood <- function(data, explained_instance, size, fixed_vari
 #'
 
 sample_locally2 <- function(data, explained_instance, explained_var, size,
-                           method = "live", fixed_variables = NULL) {
+                           method = "live", fixed_variables = NULL, ...) {
   check_conditions(data, explained_instance, size)
   explained_var_col <- which(colnames(data) == explained_var)
   if(method == "live") {
     similar <- generate_neighbourhood2(data[, -explained_var_col],
                                       explained_instance[, -explained_var_col], size,
-                                      fixed_variables)  
-  } else {
+                                      fixed_variables)
+  } else if(method == "lime") {
     similar <- permutation_neighbourhood(data[, -explained_var_col],
                                          explained_instance[, -explained_var_col],
                                          size,
                                          fixed_variables)
+  } else {
+    similar <- normal_neighbourhood(data[, -explained_var_col],
+                                    explained_instance[, -explained_var_col],
+                                    size,
+                                    fixed_variables,
+                                    ...)
   }
-  
-  explorer <- list(data = similar, 
+
+  explorer <- list(data = similar,
        target = explained_var,
        explained_instance = explained_instance,
        sampling_method = method,
@@ -115,17 +80,17 @@ sample_locally2 <- function(data, explained_instance, explained_var, size,
 #' @return A list that consists of black box model object and predictions.
 #'
 
-give_predictions2 <- function(data, black_box, explained_var, similar, predict_function, 
+give_predictions2 <- function(data, black_box, explained_var, similar, predict_function,
                              hyperpars = list(), ...) {
   if(is.character(black_box)) {
-    mlr_task <- create_task2(black_box, as.data.frame(data), explained_var)
+    mlr_task <- create_task(black_box, as.data.frame(data), explained_var)
     lrn <- mlr::makeLearner(black_box, par.vals = hyperpars)
     trained <- mlr::train(lrn, mlr_task)
     pred <- predict(trained, newdata = as.data.frame(similar))
-    list(model = mlr::getLearnerModel(trained), 
+    list(model = mlr::getLearnerModel(trained),
          predictions = pred[["data"]][["response"]])
   } else {
-    list(model = black_box, 
+    list(model = black_box,
          predictions = predict_function(black_box, similar, ...))
   }
 }
@@ -133,10 +98,10 @@ give_predictions2 <- function(data, black_box, explained_var, similar, predict_f
 
 #' Add black box predictions to generated dataset
 #'
-#' @param to_explain List return by sample_locally function.
+#' @param to_explain List return by sample_locally2 function.
 #' @param black_box_model String with mlr signature of a learner or a model with predict interface.
-#' @param data Original data frame used to generate new dataset. 
-#'        Need not be provided when a trained model is passed in 
+#' @param data Original data frame used to generate new dataset.
+#'        Need not be provided when a trained model is passed in
 #'        black_box_model argument.
 #' @param predict_fun Either a "predict" function that returns a vector of the
 #'        same type as response or custom function that takes a model as a first argument,
@@ -147,7 +112,7 @@ give_predictions2 <- function(data, black_box, explained_var, similar, predict_f
 #' @param ... Additional parameters to be passed to predict function.
 #'
 #' @return list consisting of
-#' \item{data}{Dataset generated by sample_locally function with response variable.}
+#' \item{data}{Dataset generated by sample_locally2 function with response variable.}
 #' \item{target}{Name of the response variable.}
 #' \item{model}{Black box model which is being explained.}
 #' \item{explained_instance}{Instance that is being explained.}
@@ -158,16 +123,16 @@ give_predictions2 <- function(data, black_box, explained_var, similar, predict_f
 #'
 #' @examples
 #' \dontrun{
-#' local_exploration1 <- add_predictions(wine, dataset_for_local_exploration,
+#' local_exploration1 <- add_predictions2(wine, dataset_for_local_exploration,
 #'                                       black_box_model = "regr.svm")
 #' # Pass trained model to the function.
 #' svm_model <- svm(quality ~., data = wine)
-#' local_exploration2 <- add_predictions(wine, dataset_for_local_exploration,
+#' local_exploration2 <- add_predictions2(wine, dataset_for_local_exploration,
 #'                                       black_box_model = svm_model)
 #' }
 #'
 
-add_predictions2 <- function(to_explain, black_box_model, data = NULL, predict_fun = predict, 
+add_predictions2 <- function(to_explain, black_box_model, data = NULL, predict_fun = predict,
                             hyperparams = list(), ...) {
   if(is.null(data) & is.character(black_box_model))
     stop("Dataset for training black box model must be provided")
@@ -179,9 +144,9 @@ add_predictions2 <- function(to_explain, black_box_model, data = NULL, predict_f
                                         hyperpars = hyperparams,
                                         ...)
   to_explain$data[[to_explain$target]] <- trained_black_box$predictions
-  
-  explorer <- list(data = to_explain$data, 
-       target = to_explain$target, 
+
+  explorer <- list(data = to_explain$data,
+       target = to_explain$target,
        model = trained_black_box$model,
        explained_instance = to_explain$explained_instance,
        sampling_method = to_explain$sampling_method,